+++ lib/src/repo.rs
@@ -37,6 +37,7 @@ use crate::op_store::{BranchTarget, OpStore, OperationId, RefTarget, WorkspaceId
 use crate::operation::Operation;
 use crate::rewrite::DescendantRebaser;
 use crate::settings::{RepoSettings, UserSettings};
+use crate::simple_op_heads_store::SimpleOpHeadsStore;
 use crate::simple_op_store::SimpleOpStore;
 use crate::store::Store;
 use crate::transaction::Transaction;
@@ -92,7 +93,7 @@ pub struct ReadonlyRepo {
     repo_path: PathBuf,
     store: Arc<Store>,
     op_store: Arc<dyn OpStore>,
-    op_heads_store: Arc<OpHeadsStore>,
+    op_heads_store: Arc<dyn OpHeadsStore>,
     operation: Operation,
     settings: RepoSettings,
     index_store: Arc<IndexStore>,
@@ -148,7 +149,7 @@ impl ReadonlyRepo {
         let operation_metadata =
             crate::transaction::create_op_metadata(user_settings, "initialize repo".to_string());
         let (op_heads_store, init_op) =
-            OpHeadsStore::init(op_heads_path, &op_store, &root_view, operation_metadata);
+            SimpleOpHeadsStore::init(op_heads_path, &op_store, &root_view, operation_metadata);
         let op_heads_store = Arc::new(op_heads_store);
 
         let index_path = repo_path.join("index");
@@ -225,7 +226,7 @@ impl ReadonlyRepo {
         &self.op_store
     }
 
-    pub fn op_heads_store(&self) -> &Arc<OpHeadsStore> {
+    pub fn op_heads_store(&self) -> &Arc<dyn OpHeadsStore> {
         &self.op_heads_store
     }
 
@@ -392,7 +393,7 @@ pub struct RepoLoader {
     repo_settings: RepoSettings,
     store: Arc<Store>,
     op_store: Arc<dyn OpStore>,
-    op_heads_store: Arc<OpHeadsStore>,
+    op_heads_store: Arc<dyn OpHeadsStore>,
     index_store: Arc<IndexStore>,
 }
 
@@ -405,7 +406,7 @@ impl RepoLoader {
         let store = Store::new(store_factories.load_backend(&repo_path.join("store")));
         let repo_settings = user_settings.with_repo(repo_path).unwrap();
         let op_store = Arc::from(store_factories.load_op_store(&repo_path.join("op_store")));
-        let op_heads_store = Arc::new(OpHeadsStore::load(repo_path.join("op_heads")));
+        let op_heads_store = Arc::new(SimpleOpHeadsStore::load(repo_path.join("op_heads")));
         let index_store = Arc::new(IndexStore::load(repo_path.join("index")));
         Self {
             repo_path: repo_path.to_path_buf(),
@@ -433,7 +434,7 @@ impl RepoLoader {
         &self.op_store
     }
 
-    pub fn op_heads_store(&self) -> &Arc<OpHeadsStore> {
+    pub fn op_heads_store(&self) -> &Arc<dyn OpHeadsStore> {
         &self.op_heads_store
     }
